+++ src/features/armor-mods/components/ArmorConfig.tsx
