+++ src/features/armor-mods/components/RequiredMod.tsx
